Move RootConsoleMenu into its own header and .cpp file.
