+++ core/AMBuilder
@@ -41,6 +41,7 @@ project.sources += [
   'ConsoleDetours.cpp',
   'vprof_tool.cpp',
   'smn_commandline.cpp',
+  'RootConsoleMenu.cpp',
 ]
 
 for sdk_name in SM.sdks:
